Voxelizer::Voxelize() shader setup and draw call voxelize.vert vertex shader probably finished (just passes on vertex data)
started working on voxelize.geom geometry shader
we want to use the GPU 2D rasterizer to help in voxelizing our triangles,
the rasterizer would produce a raster of fragments in the view direction,
but we want a raster in the directions of an axis aligned 3D voxel grid
so before the rasterization, we use geometry shader to project triangles
along axis that maximizes the area of the triangle 2D projection,
to minimize information loss.

after rasterization, fragment shader computes 3D voxel data
from all triangle fragments in the voxel (not yet implemented).